Moved FHIRExtractors and removed unused imports.
